# openapi.yaml
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
paths:
  /api/v1/product/:
    get:
      operationId: product_list
      tags:
      - product
      security:
      - tokenAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
          description: ''
    post:
      operationId: product_create
      tags:
      - product
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductValidate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProductValidate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductValidate'
        required: true
      security:
      - tokenAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductValidate'
          description: ''
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        category:
          type: integer
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
      required:
      - category
      - id
      - price
      - title
...
    ProductValidate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: integer
        category:
          type: integer
      required:
      - category
      - price
      - title
...
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"